@@ -1,11 +1,10 @@
 name: floating_action_row
 description: A Flutter package that provides a widget with a row or column of floating action buttons.
-version: 0.1.0
-author: Niklas Sauter <niklas.sauter@gmx.de>
-homepage: https://github.com/niklas-8/floating_action_row
+version: 0.1.1
+homepage: https://github.com/doitjustdidit/floating_action_row
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
   flutter:
