@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 
 /// Creates a divider for [FloatingActionRow]
 class FloatingActionRowDivider extends StatelessWidget {
-  final Color color;
+  final Color? color;
   final double width;
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   FloatingActionRowDivider({
     this.color,
@@ -16,7 +16,9 @@ class FloatingActionRowDivider extends StatelessWidget {
   Widget build(BuildContext context) {
     Color defaultColor =
         Theme.of(context).floatingActionButtonTheme.foregroundColor ??
-            Theme.of(context).accentIconTheme.color;
+            (Theme.of(context).brightness == Brightness.dark
+                ? Colors.black
+                : Colors.white);
 
     Widget result;
 
